requirements.txt
fastapi[all]
uvicorn
transformers
torch
pypdf
python-docx